# src/main/java/co/edu/uniquindio/sistematriage/dto/request/AtencionInputDTO.java
package co.edu.uniquindio.sistematriage.dto.request;

import co.edu.uniquindio.sistematriage.domain.enums.Estado;
import jakarta.validation.constraints.NotNull;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class AtencionInputDTO {
    @NotNull
    private Estado nuevoEstado;

    private String observacion;

}