# src/main/java/co/edu/uniquindio/sistematriage/dto/response/SugerenciaIAOutputDTO.java
package co.edu.uniquindio.sistematriage.dto.response;

import co.edu.uniquindio.sistematriage.domain.enums.Prioridad;
import co.edu.uniquindio.sistematriage.domain.enums.TipoSolicitud;
import lombok.Builder;
import lombok.Data;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Data
@Builder
public class SugerenciaIAOutputDTO {
    private TipoSolicitud tipoSugerido;
    private Prioridad prioridadSugerida;

}